backend/karmabackend/backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.3            # via django
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
coverage==5.0.3           # via -r requirements.in
django-oauth-toolkit==1.2.0  # via -r requirements.in
django-rest-auth==0.9.5   # via -r requirements.in
django==3.0.3             # via -r requirements.in, django-oauth-toolkit, django-rest-auth, djangorestframework
djangorestframework-jwt==1.11.0  # via -r requirements.in
djangorestframework==3.11.0  # via -r requirements.in, django-rest-auth
idna==2.9                 # via requests
oauthlib==3.1.0           # via django-oauth-toolkit
pip-tools==4.5.1          # via -r requirements.in
psycopg2==2.8.4           # via -r requirements.in
pyjwt==1.7.1              # via djangorestframework-jwt
pytz==2019.3              # via django
requests==2.23.0          # via django-oauth-toolkit
six==1.14.0               # via django-rest-auth, pip-tools
sqlparse==0.3.1           # via django
urllib3==1.25.8           # via requests
